backend/requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
Django==5.1
django-cors-headers==4.4.0
djangorestframework==3.15.2
pillow==10.4.0
psycopg2==2.9.9
soupsieve==2.6
sqlparse==0.5.1
tzdata==2024.1
stripe==10.10.0
paypalrestsdk==1.13.3
